+++ matjalalzz/src/main/java/shop/matjalalzz/chat/entity/MessageType.java
@@ -5,4 +5,6 @@ public enum MessageType {
     JOIN,
     LEAVE,
     KICK,
+    REQUEST_PAYMENT,
+    COMPLETE_PAYMENT
 }
